requirements.txt
Django==3.0.2
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
dj-database-url==0.5.0
python-dotenv==0.10.5
django-heroku==0.3.1
psycopg2==2.8.4
coverage==5.0.3
faker==4.0.0
factory-boy==2.12.0
africastalking==1.1.9
